requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
howlongtobeatpy>=1.0.16
